+++ src/game/neural/features.hpp
@@ -9,8 +9,8 @@ class Board;  // Forward declare
 namespace Neural {
   // roughly the idea here is to encode the boart state into a vector for features for each color
   inline constexpr size_t N_FEATURES = 64 * 6; // a bitboard for each piece type
-  typedef Vector<int32_t, N_FEATURES> FeatureVector;
-  typedef SparseVector<int32_t, N_FEATURES> FeatureDiff;
+  typedef Vector<int16_t, N_FEATURES> FeatureVector;
+  typedef SparseVector<int16_t, N_FEATURES> FeatureDiff;
 
   // We can keep these functions pure for now. So let's do that.
